requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
diskcache==5.6.3
appwrite==4.1.0
openai==1.3.7
google-generativeai==0.3.1
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
slowapi==0.1.9
